#33 wip split constants directory
